Implement N2Tiled algorithm with SoA on tiles
Implementation of a version of the N2Tiled algorithm used by fastjet,
but using a SoA structure for tiled jets which form part of the
structure of the tile itself.

LorentzVector code is used instead of a "plain" momentum vector.

Coordinate pair structure is added that tracks neighbours and
caches are used for the rightmost and the all neighbours
lists for each tile.

Some optimisations and sizehints are used to speed the code up.

Switched to plain JSON package as it's much simpler than JSON3.

Added tests against the FastJet outputs for this tiled algorithm and
for the N2Plain algorithm. An ENUM is used to control the switching
between algorithms.

This implementation remains quite slow as it suffers from many
bookkeeping overheads, unfortunately.