Implement N2Tiled algorithm using Global SoA
This is a implementation of the N2Tiled algorithm that manages
the jets as a global SoA structure and keeps a linked list of
the jets in each tile with a first -> next/previous structure.
Set() was also tried, but it was just too slow.

As jets are merged and finalised, there is shuffle and shrink
on the SoA to keep a compact list (which is quite expensive,
it turns out).

Considerable optimisation of searches using LoopVectorisation.

Finally, this is a lot faster than the per-tile SoA, but it's still
slower than the FastJet linked list approach by about x2. This slowdown
actually comes inherently from the complexity of managing the complex
SoA, which outweighs the places where it gives a performance gain!