Implement the FastJet-like N2Tiled algorithm
This is a reimplementation of the code developed by
Philippe Gras to test a Julia implementation of the FastJet
N2Tiled algorithm.

This implementation doesn't attempt any SoA, but it's extremely
lightweight in manipulations as jets are merged and finalised.

The speed is equal to or faster than FastJet, mainly down to using
the LoopVectorisation package to improve the speed of the search
for the minimum dij.