Pre-release cleanup
Write a proper README file that describes how to use the package
(real Julia documentation to come, but it's a start)

Rename main algorithms to be more consistent, viz.
sequential_jet_reconstruct and tiled_jet_reconstruct

Also remove all of the pre-cooked versions with specific values of p
(like cambridge_aachen_algo) as these add little value and would need to
exist for all algorithms.

Rename the chep.jl script to jetreco.jl and put it into an examples
directory