Refectorise plain algorithm to return ClusterSequence
We need to return the same object from all algorithms,
which is the ClusterSequence.
This allows us to implement post-facto filtering, pruning,
cleaning, etc. steps after
the initial reconstruction is done.

Explicitly convert to use PseudoJet internally, as the ClusterSequence stores these objects.
Internal algorithm is now renamed with _ as it should not be used by users.
Maintain compact arrays for the reconstruction process, but not for the jets themselves -
these are stored into the ClusterSequence linearly, keeping an
index pointer (clusterseq_index) to the correct position from the compact array.

History of recombination is filled appropriately for both
cluster-cluster merges and for
beam (final) merge steps.