Cluster sequence utilities and reconstruction animation (#65)
* Add ranking of jets

Utility function that returns a ranking of initial jets according to some
value, by default p_T

Used to assign stable values to subsequently reconstructed jets based on
their constituents, e.g., for mapping to colours for plotting

* Add intermediate state and plotting

Add a method to retrieve the intermediate state of a reconstruction
(using a special struct, JetWithAncestors)

Plotter for this intermediate state, which preserves colours based on
a ranking of primary clusters (defaults to p_T ranking)

* Fix bug in reco_state

Use the jetp_index of the parent's history entry, instead of the history
entry index (which is wrong)

Add merge_steps method to count the number of meaningful
iterations in the reconstruction
process (used to get the number
of animation steps)

* Add reconstruction animation function

animatereco() calculates all of the intermediate states during the
reconstruction and plots them with meshplot, creating an
animation output

* Add feature to plot ancestors

Option to plot all ancestors of a growing jet, using the colour of
the jet they merge to

* Shifting perspective

Allow passing of tuple of two numbers for axis view options
that give the starting and ending points for the axis view

Code refomatted

* Wrap colormap values

When there are > length(colormap) starting clusters
the extra ones are clipped, so instead wrap back to the start
of the colormap

N.B. There is an assumption here that the colormap has 256
categorical values, which may not be true. However,
I can't currently find an easy way to inspect a colormap Symbol
and extract the length

* Use variable for colormap end

* Add title and end frames options

Title does what it says on the tin...
End frames extend the animation at the end

* Minor updates

Ignore video files in the example directory
Remove unneeded dependencies from jetreco.jl

* Improved documentation

Add Documenter setup to use @autodocs to generate
all public and internal documentation.
Use @ref links to refer to methods and types in
main documentation pages.

Move get_all_ancestors() method to ClusterSequence
(it is not visualisation).

Add logo for the package!