Add constituent function
Specialise the selector return types, allowing them to take a parameter
that defines the return type (in particular PseudoJet).
Note that only LorentzVectorCyl and PseudoJet are supported.

Add a constituents() function that allows retrieval of the original clusters (as PseudoJets)
that formed a particular jet.

Add two example showing how to use this.