Support generalised kT for pp
Add a new algorithm type, GenKt, which takes an arbitrary power value.

For both strategies, add the algorithm named parameter to the argument
list, allowing the use of algorithms where both must be specified.

Add functions that will check the consistency of the power value and
algorithm type, but also set them when one or other is not specified.
Add a new list of algorithms where the power is arbitrary.

Add unittests for the consistency checks.