Support for Generalised kT pp algorithm
Reconstruction mode of GenKt now supported, when the power
value is not in [-1, 0, 1]. This is not often used, but it is a valid
algorithm choice.

For both concrete strategies and 'Best', the algorithm can now
be specified directly. If the algorithm and the power are specified
then these are checked to be consistent. If only one is specified
then the other is set to the correct value. Note that for GenKt
both are required!

ClusterSeq is adapted to record both the algorithm and the power.
Power is stored always as a Float to ensure type stability. The
constructor now needs both power and algorithm passed to it
(but this is an internal API).

set_algorithm_power_consistency! is changed to
get_algorithm_power_consistency, which returns a tuple of the
consistent power/algorithm values (can't mutate primitive types).