Generalised kT Algorithm for pp collisions (#71)
Add proper support for the generalised kT algorithm for pp collisions (i.e., operating in $(y, \phi)$ space).

Added a `JetAlgorithm` type for generalised kT, viz. `GenKt`.

Strategies are restructured to be able to be driven by only the power, or the algorithm, or (for `GenKt`) both. A new function, `get_algorithm_power_consistency` is added that will return the "partner" of a (power/algorithm) given an (algorithm/power). If inconsistent input is given an exception is thrown.

Documentation is updated explaining how to call with either `p` or `algorithm`, or both for `GenKt`.

ClusterSequence now stores the power value and the algorithm and has to be initialised with both. In addition the power is always stored as a `Float64`.

Examples are restructured to also be able to be driven by power, or algorithm, or both. However, if neither is given then the scripts will use `AntiKt` and print a warning. There is also now a clearer separation between calling the jet reconstruction itself and getting the inclusive/exclusive constituents.

A test to compare Generalised kT output with FastJet for the case `p=1.5` is added.

Tests have been added to run our code examples, to check none of these break (at least for the standalone Julia scripts), but these are currently disabled as some non-trivial problems arose:

- The dependencies are heavy and some for GLMakie actually seem to timeout
- As these are CI machines, they can't correctly setup GLMakie
- There seems to be a weird issue with setting up a different environment for the Julia nightly:

> ERROR: LoadError: ArgumentError: Package Pkg does not have Random in its dependencies

# Misc

- Notebooks are now *not* formatted by JuliaFormatter.
- For the tiled algorithm directly calling the real work interface is now discouraged (method is now `_tiled_jet_reconstruct`) to ensure that power and algorithm are handled properly in the public interface (`tiled_jet_reconstruct`). (N.B. this costs about 1us per event)

Closes #49