Implementation of Durham e+e- algorithm
The Durham (ee_kt_algorithm) is implemented:

get_algorithm_power_consistency will set p=1 in this case automatically
(this means that other power values are basically overwritten for this algorithm).

Two new functions is_pp() and is_ee() take an algorithm and return true
if the algorithm is for pp or e+e- collisions.

A new struct, EEjet, is introduced, that shares the same abstract type as PseudoJet,
but is optimised for geometric distance measures (in particular the normalised
momentum is cached, as the dot product is used to calculate the angle
between two jets). Various accessors are added for components of
the struct, as well as derived values, such as phi and rapidity.

initial_history() is moved to the ClusterSequence.jl file, which is where
it should be (used by all algorithms/strategies).
ClusterSequence structs are updated to take jets as vectors of
the abstract type FourMomentum.

The Durham EE algorithm is implemented, following much the same structure
as N2Plain. The main jet variables are stored in the cluster sequence, with
compact arrays used to calculate metric distances (with an array mapping from
the compact array back into the jet Vector).

For the generic jet_reconstruct() documentation is updated and the algorithm
handling logic is changed to support the e+e- algorithms.
If driven only by power, a pp algorithm is now explicitly assumed.

Note: as yet the outputs are not validated against fastjet.