Durham algorithm (#73)
Implementation of the e+e- k_T algorithm, aka Durham algorithm.

This is implemented in a new specific implementation as the distance metrics are rather different from those used for pp events. In particular, the reconstruction happens in (theta, phi) space.

A new EEjet structure is used, that stores the 4-momentum, but with different internal cache variables for optimising the distance calculations. (In particular the normalised momentum is cached, as the dot product is used to calculate the angle between two jets). Various accessors are added for components of the struct, as well as derived values, such as phi and rapidity.

initial_history() is moved to the ClusterSequence.jl file, which is where it should be (used by all algorithms/strategies).
ClusterSequence structs are updated to take jets as vectors of the abstract type FourMomentum.

The algorithm stores the vector of EEjets in the ClusterSequence as before (original and all intermediate jets) and optimised compact arrays for finding the closest jets are used.

Currently the implementation is about x3 slower than FastJet (which doesn't use any compact arrays). Timing on M2Pro for the ee->H input file per event is:

- JetReconstruction.jl Durham 59us
- FastJet Durham 21us
- JetReconstruction.jl AntiKt/N2Plain 33us

The consistency handling of p and algorithm is updated to take this new algorithm into account and new functions are added that map algorithms to either pp or e+e- cases (is_pp() and is_ee()).

Documentation is updated to favour passing the algorithm to the reconstruction instead of implying it from the p value (though that mode is still supported and will imply pp reconstruction).

There is also a major refactoring of tests to make testing Julia reconstruction against FastJet much easier and more independent of assumptions about the p <-> alg mapping. A new test file with e+e- -> H events is added and the 13TeV pp event file is renamed.