Add beam merge logic
It is sufficient to test for a beam merge only at the point when
the lowest dij is found, which keeps the loop logic for jet distances and
dijs much cleaner.
There is also the case of the final jet, which is handled as a beam merge
regardless of the algorithm.