Generalised kT algorithm for e+e- (#75)
Add the generalised kT algorithm for e+e-

Modifications to the _ee_genkt_algorithm() algorithm implementation 
to adapt to the generalised kT algorithm.

Modify ClusterSequence to also store directly the algorithm (which 
was missing and is needed for a full specification of the reconstruction)
Pass the ClusterSequence as an argument to the functions performing 
the reconstruction steps
- Not currently used, but should be during optimisation
Add a dij_correct_for_beam! function to check for situations where a 
particle is closer to the "beam" for the EEKt algorithm
- nni will be set to 0 in this case

Add tests for the EEKt algorithm with different values of p, R, both 
for exclusive and inclusive cases.

Add a --dump-clusterseq option to dump the cluster sequence 
after reconstruction (useful for debugging and comparing with FastJet)

Update documentation to describe also this algorithm.

Update documentation configuration to deploy tagged and 
dev documentation.

Small fix to EEJet to use cached normalised momentum vector values.

Note on performance: this is essentially the same as the Durham algorithm, 
which was noted to be currently ~x2 slower than FastJet. This is still pretty 
quick, and optimisation can now be done from the state of the algorithm working correctly.