Change input particle type to 1D AbstractArray
There is no need that the input particles be explicitly a vector, in fact
from EDM4hep files one gets a StructArrays.StructVector, which is a subtype
of AbstractArray.

This allows jet reconstruction directly from the ReconstructedParticles object
of EDM4hep, without needing to copy the data into another container.