Fix default value for p in jet_reconstruct()
The default value was wrongly set to -1, meaning that passing an algorithm
which required p!=-1 caused an algorithm/power inconsistency
when the power was not explicit.
Now default set to nothing.

Add unittests for this, including cases where things should throw.

Refactor the alg/power consistency checks to their own file, as done for
all other tests.

Expand the example test.sh to more cases (however, these did not trigger
the bug, as then the value p=nothing was explicitly passed).