Rationalise jet constructors and add preprocessing
Jet constructors are rationalised to a more standard interface.
cluster_hist_index is now always a named parameter, with a default
of 0 (in which case the jet cannot be used in a reconstruction
sequence).

Constructors are provided explicitly for making jet structs from
LorentzVector and LorentzVectorHEP, as well as a generic fallback to
any object that can supply px(), py(), pz() and energy() methods.

To properly support alternative recombination schemes we require two
functions: one is recombine, which adds two jets together; the other
is preprocess, which is called on each input particle in turn and
returns a new jet which respects the recombination scheme. In particular
making all jets massless for the pt and pt2 schemes. These parameters
are passed down through the chain from jet_reconstruct() to the actual
algorithm implementations.