Update input particles to be AbstractVector (#149)
* Update input particles to be AbstractVector

Take an AbstractVector for the algorithm entry points, which allows EDM4hep data
to be used correctly again.

* Allow AbstractVector for backend algorithms

To avoid the situation where the input particle collection is
copied from an AbstractVector, relax the stipulation
that the algorithm imlemenation must be a Vector.

For consistency, the main user entry point, jet_reconstruct()
also should take an AbstractVector as input.