Enable CLI for recombination scheme
Setup an EnumX that encodes all of the recombination schemes that
are built into the package. Map each of the enum values to a named
tuple of the appropriate "recombine" and "preprocess" methods.

In instrumented-jetreco.jl have a new argument that maps to the enum
type and use this key to the dictionary to get the named tuple of the
different methods. Then, for the actual jet reconstruction, splat this
tuple into the correct arguments.