Add contribution guide (#158)
* Add contribution guide

Add a contribution guide for the project. This is written in the
package documentation, but is linked from the standard CONTRIBUTING.md.
Added style guid and performance tips, as well as avoiding some
errors we have commonly seen.

Specify explicitly that the Julia Community Code of Conduct
should be adhered to.

Enforcement mechanism needs to be decided upon.