Improved softkiller examples
Rename --maxevents and --skip to --pileup-maxevents and --pileup-skip
which allow variations in the amount of pileup to be applied.

Select the hard scatter event with --eventno (this only even one event).

Change hard scatter and pileup files to be positional arguments, which
is more consistent with other scripts.

Make both scripts a bit more verbose for the user's information.

Run a jet reconstruction in softkiller_runtime.jl so that we actually
do something with the reduced event.

N.B. for now, the reduced event has to be rewritten into a new vector
to get the cluster_hist_indexes correct. This can be removed once
softkiller returns a good-for-reconstruction output.

Add a test.sh script that can be used for a quick test of the examples.