Internal function improvements (#195)
* Remove extra particle copy

This was an unnecessary extra copy that was being made (legacy from
when we allowed the internal method to be called directly).

* More consistent internal reconstruction interfaces

Rename all of the internal reconstruction interfaces to ! functions as
they mutate the input particle vector.

Make _ee_genkt_algorithm! take particles as a positional parameter
for consistency reasons with the other functions.

Improve docstrings.

* Format fixes

* Improved documentation

Co-authored-by: Mateusz Jakub Fila <37295697+m-fila@users.noreply.github.com>

* Additional docstring fixes

---------

Co-authored-by: Mateusz Jakub Fila <37295697+m-fila@users.noreply.github.com>