Fix return values from jet selectors
The default return values from exclusive and inclusive jet selectors
were wrong, as the LorentzVectorCyl was created mistakenly with
rapidity(jet) instead of eta(jet). This went unnoticed as the FinalJets
reversed the mistake, treating eta() as if it was rapidity().

In testing, there is quite a precision loss involved in conversion to a
LorentzVectorCyl as this involves converting from (px,py,py,E) to
(pT,eta,phi,m), then re-extracting (rapidity,phi,pT). It is the
rapidity<->phi which is problematic. So, the default return type has
been changed to be a LorentzVector as this matches better the internal
representation of PseudoJet and EEJet.

Also now import eta() and pt() methods for LorentzVectorCyl and
LorentzVector.

Decrease the tolerance on the pT test from 1e-6 to 1e-7 (for Float64).