Rationalise final jets (#198)
* First attempt at parametrising final_jets

* Fix return values from jet selectors

The default return values from exclusive and inclusive jet selectors
were wrong, as the LorentzVectorCyl was created mistakenly with
rapidity(jet) instead of eta(jet). This went unnoticed as the FinalJets
reversed the mistake, treating eta() as if it was rapidity().

In testing, there is quite a precision loss involved in conversion to a
LorentzVectorCyl as this involves converting from (px,py,py,E) to
(pT,eta,phi,m), then re-extracting (rapidity,phi,pT). It is the
rapidity<->phi which is problematic. So, the default return type has
been changed to be a LorentzVector as this matches better the internal
representation of PseudoJet and EEJet.

Also now import eta() and pt() methods for LorentzVectorCyl and
LorentzVector.

Decrease the tolerance on the pT test from 1e-6 to 1e-7 (for Float64).

* Formatting

* Add jet selector tests

Test if different struct selectors work for selected jets, both
inclusive and exclusive.

Add mass and mass2 "imports" for LorentzVector and LorentzVectorHEP.

* Broaden test coverage

* Add new tests to runtests.jl

* Format!

* Test mass and eta as well

* Better isapprox for FinalJet

Pass kwargs properly and don't rely on the ≈ alias.

Use the default equality implementation.